online_inference/requirements.txt
antlr4-python3-runtime==4.8
astroid==2.5.6
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Faker==8.1.4
fastapi==0.65.1
h11==0.12.0
hydra-core==1.0.6
idna==2.10
iniconfig==1.1.1
isort==5.8.0
joblib==1.0.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
numpy==1.20.3
omegaconf==2.0.6
packaging==20.9
pluggy==0.13.1
py==1.10.0
pydantic==1.8.2
pylint==2.8.2
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sklearn==0.0
starlette==0.14.2
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
uvicorn==0.13.4
wrapt==1.12.1